Extract `version.js` into different classes and extend tests
phovea/generator-phovea#425
